README.md
This app has php functionality and mysql data storing. It allows the user to insert his or her name, surname, date of booking in and out of the hotel and which hotel they would prefer to stay at. Then process the amount of days the will stay and how much it will cost. They then can decide to book the hotel and the information that they provided will then be stored in a database with the name of Hotel.